Better network retrying (so you don't have to restart just because the vantage server wasn't available on the network for a while).
Handle IRZone Buttons.

New Shade3 blind type that combines 2 or 3 drycontacts and a sensor to make a fully-controllable shade.  They're named with "open", "close", "stop" (optional) and the sensor is "is open" suffixes.

Bumping to 0.0.39 but this really hasn't been tested on my system yet.  I want to be sure it deploys all the way through correctly since I'm now using a completely different deployment setup for running hass.